pyproject.toml
[project]
name = "rag-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "mcp>=1.26.0",
    "sentence-transformers>=5.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "pymupdf>=1.27.2",
]

[project.scripts]
rag-mcp = "main:mcp.run"
